vk: begin texture r_/vk_ split
no functional changes, just copied some functions over
