+++ ref/vk/infotool.c
@@ -67,12 +67,12 @@ void XVK_CameraDebugPrintCenterEntity( void ) {
 		const int surface_index = surf - ent->model->surfaces;
 		const texture_t *current_tex = R_TextureAnimation(ent, surf, NULL);
 		const int tex_id = current_tex->gl_texturenum;
-		const vk_texture_t* const texture = R_TextureGetByIndex( tex_id );
+		const char *const tex_name = R_TextureGetNameByIndex( tex_id );
 		const texture_t *tex = surf->texinfo->texture;
 
 		p += Q_snprintf(p, end - p,
 			"surface index: %d; texture: %s(%d)\n",
-			surface_index, texture ? texture->name : "NONE", tex_id
+			surface_index, tex_name ? tex_name : "NONE", tex_id
 		);
 
 		if (tex->anim_total > 0 && tex->anim_next) {
@@ -80,9 +80,9 @@ void XVK_CameraDebugPrintCenterEntity( void ) {
 			p += Q_snprintf(p, end - p,
 				"anim textures chain (%d):\n", tex->anim_total);
 			for (int i = 0; i < tex->anim_total && tex; ++i) {
-				const vk_texture_t *vkt = R_TextureGetByIndex(tex->gl_texturenum);
+				const char* const texname = R_TextureGetNameByIndex(tex->gl_texturenum);
 				p += Q_snprintf(p, end - p,
-					"%d: %s(%d)%s\n", i, vkt ? vkt->name : "NONE", tex->gl_texturenum, tex == current_tex ? " <-" : "   ");
+					"%d: %s(%d)%s\n", i, texname ? texname : "NONE", tex->gl_texturenum, tex == current_tex ? " <-" : "   ");
 				tex = tex->anim_next;
 			}
 		}
