+++ ref/vk/vk_lightmap.c
@@ -227,7 +227,7 @@ void VK_UploadLightmap( void )
 void VK_ClearLightmap( void )
 {
 	for (int i = 0; i < gl_lms.current_lightmap_texture; ++i)
-		R_TextureRelease(tglob.lightmapTextures[i]);
+		R_TextureFree(tglob.lightmapTextures[i]);
 	gl_lms.current_lightmap_texture = 0;
 
 	LM_InitBlock();
