+++ ref/vk/vk_overlay.c
@@ -291,7 +291,7 @@ void R_DrawStretchRaw( float x, float y, float w, float h, int cols, int rows, c
 
 void R_DrawTileClear( int texnum, int x, int y, int w, int h )
 {
-	PRINT_NOT_IMPLEMENTED_ARGS("%s", R_TextureGetByIndex(texnum)->name );
+	PRINT_NOT_IMPLEMENTED_ARGS("%s", R_TextureGetNameByIndex(texnum));
 }
 
 void CL_FillRGBA( float x, float y, float w, float h, int r, int g, int b, int a )
